Bundles no longer require config.yaml to execute
The metaprogramming we were using to set up the code was requiring a
`config.yaml` file be present at execution time. However, this was only
being used to get the names of the bundle's commands; it would use this
to then require all the appropriate command implementation files. In
reality, we only really need to load the one file corresponding to the
command we're currently executing.

This change now only does the metaprogramming on the current command's
implementation, eliminating the need for there to be a `config.yaml`.

(Long term, we'll remove the metaprogramming altogether, but this is a
simplifying step in that direction.)